+++ v3.0/src/__init__.py
@@ -1,0 +1,1 @@
+"""StockPredictor V3.0 source package."""